+++ lib/src/db/mis/sd_arrawdsum.dart
@@ -80,7 +80,7 @@ class SdArrAwdSumArchive extends mis.MisReportArchive {
       await dbConfig.coll.insertAll(data);
       print(
           '--->  Inserted $reportName for account $account, month $date, version $version, tab $tab successfully');
-      return Future.value(0);
+      return await Future.value(0);
     } catch (e) {
       print('XXX $e');
       return Future.value(1);