+++ lib/src/db/mis/tr_sch3p2.dart
@@ -87,7 +87,7 @@ class TrSch3p2Archive extends mis.MisReportArchive {
       await dbConfig.coll.insertAll(data);
       print(
           '--->  Inserted $reportName for account $account, month $month, version $version, tab $tab successfully');
-      return Future.value(0);
+      return await Future.value(0);
     } catch (e) {
       print('XXX $e');
       return Future.value(1);