+++ sql/hq/hq_hydro_data.sql
@@ -1,3 +1,4 @@
+load ggsql;
 
 SELECT * FROM WaterLevel;
 
@@ -10,6 +11,66 @@ ORDER BY date;
 SELECT DISTINCT station_id
 FROM WaterLevel
 ORDER BY station_id;
+
+
+-- One station hourly data
+SELECT hour_beginning, value
+FROM WaterLevel
+-- WHERE station_id = '1-2944'
+-- WHERE station_id = '1-2987'
+-- WHERE station_id = '1-3009'
+WHERE station_id = '1-2919'
+-- AND value > 300
+ORDER BY hour_beginning
+VISUALIZE 
+    hour_beginning as x,
+    value as y
+DRAW line
+;
+
+--- Get the median value by day
+SELECT hour_beginning::DATE AS date, median(value) AS value
+FROM WaterLevel
+WHERE station_id = '1-2919'
+GROUP BY date
+ORDER BY date
+VISUALIZE 
+    date as x,
+    value as y
+DRAW line
+;
+
+
+
+--- Aggregate all stations (there is some bad data)
+SELECT hour_beginning, SUM(value) AS total_value
+FROM WaterLevel
+GROUP BY hour_beginning
+ORDER BY hour_beginning
+VISUALIZE 
+    hour_beginning as x,
+    total_value as y
+DRAW line
+;
+
+--- Plot all stations as separate lines
+SELECT station_id, hour_beginning::DATE AS date, median(value) AS value
+FROM WaterLevel
+WHERE value > 0
+GROUP BY station_id, date
+ORDER BY date
+VISUALIZE 
+    date as x,
+    value as y,
+    station_id as color
+DRAW line
+SCALE ORDINAL color
+;
+
+
+
+
+
 
 SELECT station_id, MIN(value) AS min_value, MAX(value) AS max_value
 FROM WaterLevel
@@ -61,7 +122,8 @@ WHERE station_id == '1-2951'
 GROUP BY station_id, date
 ORDER BY date;
 
-
+-- ============================================================================================================
+-- Read